Add route file differently
